requirements.txt
markdown~=3.3.3
mako~=1.1.4
Pillow~=8.1.0
git+https://github.com/hbldh/hitherdither
pyyaml~=5.4.1
git+https://github.com/pygments/pygments
html5print
lxml
numpy
scipy
sklearn
mistune==2.0.0
pyfiglet